Rename submodule `region_markers.py` -> `markers`
